requirements.txt
numpy==1.26.4
pillow==11.0.0
pandas==2.2.3
scikit-learn==1.7.2
scipy==1.16.2
matplotlib==3.10.6
tqdm==4.67.1

# deep learning
torch==2.2.2
torchvision==0.17.2
facenet-pytorch==2.6.0

# image processing
opencv-python==4.10.0.84
scikit-image==0.22.0

# about
python-dotenv==1.1.1
joblib==1.5.2